raspberry-pi/requirements.txt
azure-servicebus>=7.11.0
azure-identity>=1.12.0
python-dotenv>=1.0.0
